refactored terminal component
