+++ client/src/app.tsx
@@ -1,17 +1,15 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import "./app.css";
 import Router from "./router";
 
 import { createTheme, ThemeProvider } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import { createBrowserHistory } from 'history'
 
 const darkTheme = createTheme({
   palette: {
     mode: "dark",
   },
 });
-
 
 function App() {
