+++ client/src/components/StreamPreview.tsx
@@ -1,10 +1,11 @@
 import { FC, ReactElement, useState, useEffect } from "react";
-import WSTerminal from "./WSTerminal";
+import Terminal from "./Terminal";
 import PubSub from "./../lib/pubsub";
 import * as utils from "../utils";
 import * as constants from "../lib/constants";
 import * as message from "../types/message";
 
+import Uptime from "./Uptime";
 import dayjs from "dayjs";
 import customParseFormat from 'dayjs/plugin/customParseFormat';
 
@@ -24,7 +25,6 @@ interface Props {
 
 const StreamPreview: FC<Props> = ({ title, wsUrl, streamerID, nViewers, startedTime, lastActiveTime }): ReactElement => {
 
-  const [ upTime, setUpTime ] = useState(utils.formatDuration(dayjs(), dayjs(startedTime)));
   const [ msgManager, setMsgManager ] = useState<PubSub>();
 
   useEffect(() => {
@@ -38,6 +38,7 @@ const StreamPreview: FC<Props> = ({ title, wsUrl, streamerID, nViewers, startedT
     utils.sendWhenConnected(ws, payload);
 
     const tempMsg = new PubSub();
+
     ws.onmessage = (ev: MessageEvent) => {
       let msg = JSON.parse(ev.data);
 
@@ -57,22 +58,19 @@ const StreamPreview: FC<Props> = ({ title, wsUrl, streamerID, nViewers, startedT
           break;
 
       }
+    }
 
-   }
     tempMsg.sub("request", (msgType: string) => {
-
       var payload = JSON.stringify({
         Type: msgType,
         Data: "",
       });
-      utils.sendWhenConnected(ws, payload);
+      setTimeout(() => {
+        ws.send(payload);
+      }, 100);
     })
 
     setMsgManager(tempMsg);
-
-    setInterval(() => {
-      setUpTime(utils.formatDuration(dayjs(), dayjs(startedTime)));
-    }, 1000);
 
     // preview doesn't need to be live
     setTimeout(() => {
@@ -81,14 +79,17 @@ const StreamPreview: FC<Props> = ({ title, wsUrl, streamerID, nViewers, startedT
 
   }, [])
 
-
   return (
     <div className="relative bg-black rounded-lg w-full my-4">
       {msgManager &&
-      <WSTerminal msgManager={msgManager} height={window.innerWidth > 600 ? 350 : 250} width={window.innerWidth > 600 ? 500 : window.innerWidth - 35 }/>
+      <Terminal 
+        className="bg-black"
+        msgManager={msgManager} 
+        height={window.innerWidth > 600 ? 350 : 250} 
+        width={window.innerWidth > 600 ? 500 : window.innerWidth - 35 }/>
       }
       <div className="p-1 bg-red-400 rounded-lg absolute top-4 right-4">
-        <p className="text-mdtext-whtie font-semibold">{upTime}</p>
+        <Uptime className="text-md text-white font-semibold" startTime={new Date(startedTime)} />
       </div>
 
       <div className="absolute bottom-0 left-0 w-full rounded-b-lg bg-gray-600 bg-opacity-90 p-4" >
